refactor(search): split sources into their own files
